Updated the automator
Have automated the workbook and scenario mapping further. it now takes a blank template and fills it in based on the filters. like before, but now for each scenario in the REPORT00 file.

Filters files now dont need WorkbookID, or scenario name columns - this is automated.

Uploaded the latest filter files from the onedrive.